# lib/api-zod/src/generated/types/createTaskInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marketing & Project Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTaskInputPriority } from "./createTaskInputPriority";
import type { CreateTaskInputStatus } from "./createTaskInputStatus";

export interface CreateTaskInput {
  title: string;
  description?: string | null;
  projectId?: number | null;
  status: CreateTaskInputStatus;
  priority: CreateTaskInputPriority;
  assignee?: string | null;
  dueDate?: string | null;
}