# lib/api-zod/src/generated/types/createTeamMemberInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marketing & Project Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTeamMemberInput {
  name: string;
  email: string;
  role: string;
  department?: string | null;
}