# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marketing & Project Management API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from "./taskPriority";
import type { TaskStatus } from "./taskStatus";

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  projectId?: number | null;
  status: TaskStatus;
  priority: TaskPriority;
  assignee?: string | null;
  dueDate?: string | null;
  createdAt: string;
}