# lib/api-zod/src/generated/types/updateCampaignInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marketing & Project Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCampaignInputStatus } from "./updateCampaignInputStatus";
import type { UpdateCampaignInputType } from "./updateCampaignInputType";

export interface UpdateCampaignInput {
  name?: string;
  description?: string | null;
  status?: UpdateCampaignInputStatus;
  type?: UpdateCampaignInputType;
  budget?: number;
  spent?: number | null;
  leads?: number | null;
  conversions?: number | null;
  startDate?: string | null;
  endDate?: string | null;
}