# lib/api-zod/src/generated/types/updateTaskInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marketing & Project Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTaskInputPriority } from "./updateTaskInputPriority";
import type { UpdateTaskInputStatus } from "./updateTaskInputStatus";

export interface UpdateTaskInput {
  title?: string;
  description?: string | null;
  projectId?: number | null;
  status?: UpdateTaskInputStatus;
  priority?: UpdateTaskInputPriority;
  assignee?: string | null;
  dueDate?: string | null;
}